+++ projects/rocprofiler-sdk/CHANGELOG.md
@@ -264,3 +264,4 @@ Full documentation for ROCprofiler-SDK is available at [rocm.docs.amd.com/projec
 ### Changed
 
 - Version updated to 1.2.0 to support better library compatibility detection for downstream dependencies
+- Fixed rocpd OTF2 output to add ACCELERATOR_DEVICE as system tree node domain for AMD devices.